Fold PositionedFauxCaret into FauxCaret component.
